+++ sei-ibc-go/modules/light-clients/09-localhost/types/client_state.go
@@ -6,8 +6,6 @@ import (
 	"reflect"
 	"strings"
 
-	ics23 "github.com/confio/ics23/go"
-
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -57,11 +55,6 @@ func (cs ClientState) Validate() error {
 	if cs.Height.RevisionHeight == 0 {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidHeight, "local revision height cannot be zero")
 	}
-	return nil
-}
-
-// GetProofSpecs returns nil since localhost does not have to verify proofs
-func (cs ClientState) GetProofSpecs() []*ics23.ProofSpec {
 	return nil
 }